Adding model classes Pizza and Customer
