--- BackEnd/src/main/java/com/revature/models/Customer.java
@@ -1,14 +1,0 @@
-package com.revature.models;
-
-
-public class Customer{
-
-    // Properties
-    private int id;
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-
-
-}
