+++ BackEnd/src/main/java/org.swagarelli/models/Customer.java
@@ -1,0 +1,29 @@
+package org.swagarelli.models;
+
+
+public class Customer{
+
+    // Properties
+    private int id;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+    private String address;
+    private boolean isMember;
+
+    // Constructors
+    // No-arg constructor
+    public Customer() {
+        id = 0;
+        firstName = "";
+        lastName = "";
+        email = "";
+        password = "";
+        address = "";
+        isMember = false;
+    }
+
+    // Overloading the constructor with the
+
+}
